refactor(dashboard): extract routing/model-table modules from FallbackPage
Pure move, no behavior change. FallbackPage was a 1000-line page that also
exported the shared table internals consumed by ModelDetailPage and the
command palette. Now:
- lib/routing.ts: FallbackEntry/RoutingData/Row types + pure helpers
  (formatTokens, quota badges, buildGroups, platformColors)
- components/model-table.tsx: AxisBar, ModelTableHead, RowContent,
  GroupHeaderCells, SortableGroupRow
- components/token-usage-bar.tsx: TokenUsageBar
- components/custom-weights-popover.tsx: CustomWeightsPopover
- FallbackPage.tsx shrinks 1032 -> ~390 lines and exports only the page;
  importers now use the shared modules
